+++ src/Digdir.Domain.Dialogporten.Application/Common/Behaviours/DomainContextBehaviour.cs
@@ -8,11 +8,11 @@ namespace Digdir.Domain.Dialogporten.Application.Common.Behaviours;
 /// <summary>
 /// This behaviour is used to ensure that the domain context is valid and return the correct response to the presentation layer if it is not.
 /// <para>
-/// The behaviour will return <see cref="DomainError"/> if <typeparamref name="TResponse"/> implements 
+/// The behaviour will return <see cref="DomainError"/> if <typeparamref name="TResponse"/> implements
 /// <see cref="OneOf.IOneOf"/> containing <see cref="DomainError"/> on invalid domain state.
 /// </para>
 /// <para>
-/// The behaviour will throw a <see cref="DomainException"/> if <typeparamref name="TResponse"/> does not 
+/// The behaviour will throw a <see cref="DomainException"/> if <typeparamref name="TResponse"/> does not
 /// implement <see cref="OneOf.IOneOf"/> containing <see cref="DomainError"/> on invalid domain state.
 /// </para>
 /// </summary>