+++ src/Digdir.Domain.Dialogporten.GraphQL/Program.cs
@@ -24,6 +24,7 @@ var telemetryConfiguration = TelemetryConfiguration.CreateDefault();
 Log.Logger = new LoggerConfiguration()
     .MinimumLevel.Warning()
     .Enrich.FromLogContext()
+    .Enrich.WithEnvironmentName()
     .WriteTo.Console(formatProvider: CultureInfo.InvariantCulture)
     .WriteTo.ApplicationInsights(telemetryConfiguration, TelemetryConverter.Traces)
     .CreateBootstrapLogger();
@@ -51,6 +52,7 @@ static void BuildAndRun(string[] args, TelemetryConfiguration telemetryConfigura
         .ReadFrom.Configuration(context.Configuration)
         .ReadFrom.Services(services)
         .Enrich.FromLogContext()
+        .Enrich.WithEnvironmentName()
         .WriteTo.Console(formatProvider: CultureInfo.InvariantCulture)
         .WriteTo.ApplicationInsights(telemetryConfiguration, TelemetryConverter.Traces));